+++ src/components/DataFetcher/index.tsx
@@ -2,10 +2,9 @@ import React from 'react';
 
 import useSWRImmutable from 'swr/immutable';
 
-import ErrorMessage from './ErrorMessage';
-
 import { fetcher } from 'src/api';
 import Spinner from 'src/components/dls/Spinner/Spinner';
+import ErrorMessage from 'src/components/Error';
 import { BaseResponse } from 'types/ApiResponses';
 
 interface Props {
